// types/UserType.ts
export type Usertype = {
    avatar_url: string | null;
    name: string | null;
    login: string | null;
    twitter_username: string | null;
    public_repos: number | null;
    followers: number | null;
    blog: string | null;
    location: string | null;
    created_at: string | null;
    company: string | null;
    bio: string | null;
    following: number | null
}